refactor(menu-bar): list renamed to submenu
