+++ src/layouts/workbench-window/application-header/menu-bar/MenuBar.tsx
@@ -1,5 +1,5 @@
-import MenuBarFileButton from "@/layouts/workbench-window/application-header/menu-bar/file-list/MenuBarFileButton";
-import MenuBarViewButton from "@/layouts/workbench-window/application-header/menu-bar/view-list/MenuBarViewButton";
+import MenuBarFileButton from "@/layouts/workbench-window/application-header/menu-bar/file-submenu/MenuBarFileButton";
+import MenuBarViewButton from "@/layouts/workbench-window/application-header/menu-bar/view-submenu/MenuBarViewButton";
 
 export default function MenuBar() {
   return (
