+++ src/layouts/workbench-window/application-header/menu-bar/view-submenu/MenuBarViewButton.tsx
